playlist feature on progress

+++ app/Http/Controllers/adminApi.php
@@ -2,7 +2,6 @@
 
 namespace App\Http\Controllers;
 use Illuminate\Http\Request;
-use App\Models\Admin;
 use App\Models\User;
 use App\Models\Channel;
 use App\Models\Video;
@@ -13,14 +12,45 @@ use Hash;
 
 class adminApi extends Controller
 {
+  
+  public function dashboard(){
+    $total_users = User::where('is_admin', 0)->count();
+    $total_admins = User::where('is_admin', 1)->count();
+    $total_videos = Video::query()->count();
+    $total_categories = Category::query()->count();
+    return [
+      'success' => true,
+      'total_users' => $total_users,
+      'total_admins' => $total_admins,
+      'total_videos' => $total_videos,
+      'total_categories' => $total_categories,
+    ];
+  }
+  
   // Get all users
   public function getUsers(){
     return User::where('is_admin', 0)->get();
   }
   
+  // Get all admins
+  public function getAdmins(){
+    return User::where('is_admin', 1)->get();
+  }
+  
   // Get all Channels
   public function getChannels(){
     return Channel::all();
+  }
+  
+  // Give a user 'admin' role
+  public function makeAdmin($id){
+    $user = User::find($id);
+    $user->is_admin = 1;
+    $user->tokens()->delete();
+    if($user->save()){
+      return ['success' => true, 'message' => 'Admin access granted to the user!'];
+    }
+    return response()->json(['success' => false, 'message' => 'failed to make admin!'], 451);
   }
   
   // Add new category of video
@@ -45,9 +75,13 @@ class adminApi extends Controller
   
   // Delete a user
   public function removeUser($id) {
-    $result = User::findOrFail($id)->delete();
-    $r2 = Video::where('uploader_id', $id)->delete();
-    $r3 = Channel::find($id)->delete();
+    $user = User::findOrFail($id);
+    if($user->is_admin){
+      return accessDenied();
+    }
+    $result = $user->delete();
+    $r2 = $user->videos->delete();
+    $r3 = $user->channel->delete();
     $r5 = Subscriber::where('channel_id', $id)->delete();
     $r6 = Notification::where('from', $id)->delete();
     if ($result){
